Add edit & save function
Use of primitive types in JTable may cause undefined behaviors. Manually
converting them into Object forms solves the issue. GUI is left for
future implementation.

+++ app/src/main/java/com/group89/app/model/MarkRecordList.java
@@ -28,9 +28,9 @@ public class MarkRecordList extends ArrayList<MarkRecord> {
   public double getAverageMark() {
     double average = 0;
     for (MarkRecord record : this) {
-      average += (double) record.getMark() * (double) record.getCredits();
+      average += (double) record.getMark() * record.getCredits();
     }
-    average /= (double) this.getTotalCredits();
+    average /= this.getTotalCredits();
     return average;
   }
 }
